accounts are being updated

+++ tenmo-server/src/main/java/com/tenmo/model/Account.java
@@ -3,11 +3,11 @@ package com.tenmo.model;
 import java.math.BigDecimal;
 
 public class Account {
-    long accountId;
-    long userId;
+    int accountId;
+    int userId;
     BigDecimal balance;
 
-    public Account(long accountId, long userId, BigDecimal balance) {
+    public Account(int accountId, int userId, BigDecimal balance) {
         this.accountId = accountId;
         this.userId = userId;
         this.balance = balance;
@@ -20,7 +20,7 @@ public class Account {
         return accountId;
     }
 
-    public void setAccountId(long accountId) {
+    public void setAccountId(int accountId) {
         this.accountId = accountId;
     }
 
@@ -28,7 +28,7 @@ public class Account {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
